% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survtime_generator.R
\name{survtime_generator}
\alias{survtime_generator}
\title{Discrete-time survival time generator}
\usage{
survtime_generator(b_haz, betas, hazard, ids, idvar, n, ncov, stime, x)
}
\arguments{
\item{b_haz}{A vector of "baseline hazards" for all patients, with length equal to 'stime'}

\item{betas}{A data.frame with length equal to the length of 'x', and width equal to (ncov + 3)}

\item{hazard}{A hazard generating function. The default is haz_func(). The function must take arguments:
t, x, betas, b_haz, and ncov. For details, see the documentation of haz_func()}

\item{ids}{Patient ids.}

\item{idvar}{The name of the variable in 'x' which holds patient ids.}

\item{n}{TODO}

\item{ncov}{The number of covariates (the number of baseline covariates plus the number of time-varying covariates)}

\item{stime}{The number of follow-up times}

\item{x}{A data.frame resulting from a call to fd_generator()}
}
\description{
Discrete-time survival time generator
}
