% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weihaz.R
\name{weihaz}
\alias{weihaz}
\title{A hazard generator for a weibull distribution}
\usage{
weihaz(x, shape, scale)
}
\arguments{
\item{x}{A vector of positive real numbers at which to evaluate the hazard}

\item{shape}{The shape parameter}

\item{scale}{The scale parameter}
}
\value{
A vector of hazards evaluated at each element of argument 'x'
}
\description{
A hazard generator for a weibull distribution
}
